+++ electron.vite.config.ts
@@ -45,7 +45,10 @@ export default defineConfig(({ mode }) => ({
   preload: {
     build: {
       rollupOptions: {
-        input: 'src/preload/index.ts'
+        input: {
+          index: 'src/preload/index.ts',
+          webviewPreload: 'src/webviewPreload/index.ts'
+        }
       },
       outDir: 'build/preload',
       minify: true,
