+++ src/backend/constants/paths.ts
@@ -46,6 +46,11 @@ export const publicDir = resolve(
   app.isPackaged || process.env.CI === 'e2e' ? '' : '../public'
 )
 
+// Built preload scripts live next to the main process bundle
+// (`build/preload/`). The webview preload is built from
+// `src/webviewPreload/index.ts` — see `electron.vite.config.ts`.
+const preloadDir = resolve(__dirname, '..', 'preload')
+
 export const fakeEpicExePath = fixAsarPath(
   join(publicDir, 'bin', 'x64', 'win32', 'EpicGamesLauncher.exe')
 )
@@ -55,7 +60,7 @@ export const galaxyCommunicationExePath = fixAsarPath(
 )
 
 export const webviewPreloadPath = fixAsarPath(
-  join('file://', publicDir, 'webviewPreload.js')
+  join('file://', preloadDir, 'webviewPreload.js')
 )
 
 /**
